requirements.txt

torch==2.3.1


fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.37.1
requests==2.32.3
starlette==0.37.2


rq==1.16.2
redis==5.0.7
watchdog==4.0.2
cloudpickle==3.0.0


SQLAlchemy==2.0.35
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1


loguru==0.7.2
typer==0.12.3
tenacity==8.5.0
tqdm==4.66.4
rich==13.7.1
pynvml==11.5.0
python-dateutil==2.9.0.post0
filetype==1.2.0
python-magic==0.4.27


PyMuPDF==1.24.9
pypdf==4.3.1
pdfplumber==0.11.4
python-docx==1.1.2
pillow==10.4.0
opencv-python==4.10.0.84
tabula-py==2.10.0
webvtt-py==0.4.6

python-doctr[torch]==0.9.0
pytesseract==0.3.10


transformers==4.43.4
accelerate==0.33.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
qdrant-client==1.10.1
openai==1.45.0
httpx==0.27.0
tiktoken==0.7.0


pandas==2.2.2
numpy==1.26.4
scipy==1.13.1

docxtpl==0.16.7
reportlab==4.2.2


pytest==8.3.2
python-multipart==0.0.9
docx2txt==0.8
nltk